orderservice/requirements.txt
Django>=4.2,<6.1
djangorestframework>=3.14

# HTTP calls (product service)
requests>=2.31

# Redis (cart)
redis>=5.0

# RabbitMQ (event system)
pika>=1.3

# Payments
stripe>=8.0

# Testing
pytest>=7.4
pytest-django>=4.7